:recycle: Move service view to `drf-standardized-errors`
